refactor(strategy): MoveStrategy와 CarMoveStrategy를 strategy 패키지로 이동
