+++ src/main/java/racingcar/strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.util.movement;
+package racingcar.strategy;
 
 public interface MoveStrategy {
     boolean canMove(int number);
